Move VTM GO libs to seperate folder
