+++ resources/lib/vtmgo/vtmgoauth.py
